requirements.txt
click>=8.0.0
rich>=10.0.0
pyyaml>=6.0.0
ruamel.yaml>=0.18.0
openai>=1.0.0
tenacity>=8.0.0
pydantic>=1.10.0
PyGithub>=1.59.0
python-dotenv>=0.19.0
google-generativeai>=0.3.2 